NET-466: when marking datum aggregates as "stale" support removed datum.
When aggregate datum are marked stale via the API, a date range is used
and that range is used to query the raw datum to generate the list of
aggregate buckets that must be recomputed in the given date range. It is
possible for datum to have been deleted, however, with aggregate buckets
previously existing that are no longer valid and should be deleted. This
change includes those aggregate buckets in the query of "stale" buckets
so they can be removed.