NET-468: include HTTP method, URI, body in error C2C events.
The original "making request" event already included the HTTP method,
URI, and request body, but if an error occurred the subsequent error
event lacked those details, making it harder to correlate to the actual
request that was made.

This change updates the error event data to include the HTTP method,
URI, and request body. Additionally the HTTP response body will be
included (up to a maximum size) if the error is a HTTP status error (for
example a 4xx or 5xx error).