NET-475: hourly cached aggs should require rows with accumulating properties.
This update fixes the hourly aggregation cache process to align with the
on-the-fly difference calculation, in terms of requiring the difference
rows contributing to the aggregation to have non-NULL data_a values.

The queries are complicated by performance constraints: the tolerance
for reading difference is 1 year, but searching through that many rows
looking for non-NULL data_a values takes far too long because comparing
on the data_a column requires a full index scan vs an index-only scan
that can perform much faster when only comparing the timestamps.

The techniques applied to achieve reasonable performance are:

1. If the stream meta has a NULL names_a value, assume all data_a values
   are NULL, and skip the search entirely.
2. Try to find a non-null data_a row within 14 days. That is a short
   enough time period that even if no value found it can complete
   quickly enough.
3. Fall back to a full 1-year search not considering the data_a column,
   to find any row within the 1-year period. If a row is found, then
   that specific row is checked for a non-NULL data_a value.

There are 3 styles of datum streams these queries are trying to serve:

 * The names_a meta is non-NULL and every row populates a data_a value.
   This is the "happy path".
 * The names_a meta is NULL and every row has a NULL data_a value. The
   optimization #1 above addresses this.
 * The names_a meta is non-NULL but every row as a NULL data_a value.
   This is a "worst case" scenario because the query would have to
   always search the full tolerance, never finding a row with a non-NULL
   data_a value. The optimization #2 addresses this by limiting the
   scope to a short enough window to allow the query to finish quickly.
 * The names_a meta is non-NULL but data_a values are populated
   intermittently. Optimizations #2 and #3 address this similarly to the
   previous point.

Longer term, if a true 1-year window is required (i.e. optimizaiton #2
should be removed) then some other technique will be required, such as
adding a (possibly partial) expression index on something like

	CARDINALITY(data_a) > 0

or adding a dervied `has_a` column and creating a new primary key index
that INCLUDES that column so index-only scans can be used.